# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/recipes": {
            "get": {
                "tags": [
                    "Recipe"
                ],
                "summary": "Список",
                "operationId": "99b2921505aaa0b05aafd140ed9f5f09",
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Recipe"
                ],
                "summary": "Создание",
                "operationId": "663b6533d0b183c33766adc163cdf85a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_type_product": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_raw_material": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "кг"
                                    },
                                    "quantity": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 9
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Описание"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/recipes/{recipe}": {
            "get": {
                "tags": [
                    "Recipe"
                ],
                "summary": "Список по ID",
                "operationId": "da0e5a21cf03163b56d6d3c1c46c65dd",
                "parameters": [
                    {
                        "name": "recipe",
                        "in": "path",
                        "description": "ID recipe",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Recipe"
                ],
                "summary": "Удаление",
                "operationId": "3d5c8451379b583594b3f680c70cd602",
                "parameters": [
                    {
                        "name": "recipe",
                        "in": "path",
                        "description": "ID recipe",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "done"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Recipe"
                ],
                "summary": "Обновление",
                "operationId": "9b8af4b7323f39b3c4a938fac74ecf3e",
                "parameters": [
                    {
                        "name": "recipe",
                        "in": "path",
                        "description": "ID recipe",
                        "required": true,
                        "example": 2
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_type_product": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_raw_material": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "кг"
                                    },
                                    "quantity": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 9
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Описание"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sklads": {
            "get": {
                "tags": [
                    "Sklad"
                ],
                "summary": "Список",
                "operationId": "b26f11a4a09ee28c180e78062aec165d",
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sklad"
                ],
                "summary": "Создание",
                "operationId": "ef1edafc0ae6dd265266c81f6b5138b1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Склад 1"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Гафуров, 13а"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Готовые продукты"
                                    },
                                    "phone": {
                                        "type": "integer",
                                        "example": 4040
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sklads/{sklad}": {
            "get": {
                "tags": [
                    "Sklad"
                ],
                "summary": "Список по ID",
                "operationId": "e48ea8268a1ccbe00ee8a9c100ccc6cb",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "ID sklad",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sklad"
                ],
                "summary": "Удаление",
                "operationId": "34f522da0ac3a81417f0faa84ee0500e",
                "parameters": [
                    {
                        "name": "sklad",
                        "in": "path",
                        "description": "ID sklad",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "done"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Sklad"
                ],
                "summary": "Обновление",
                "operationId": "16bc8f06993d9ab061b722b38057f70f",
                "parameters": [
                    {
                        "name": "sklad",
                        "in": "path",
                        "description": "ID sklad",
                        "required": true,
                        "example": 2
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Склад 1"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Гафуров, 13а"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Готовые продукты"
                                    },
                                    "phone": {
                                        "type": "integer",
                                        "example": 4040
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}